[Tenable.sc BUG] Exhibiting non-well-formed XML response as failed test
Note: This is not a bug in pyTenable, but simply an exhibition of a problem in
Tenable.sc.

It appears that Tenable.SecurityCenter 5.20.0 introduces an issue wherein
the returned XML within the "pluginText" JSON node no longer XML-encodes
special characters such as > and <. As a result, any <plugin_output/> documents
that happen to include > or < cause issues.

As shown in pyTenable test cases, a recent response from
Tenable.SecurityCenter 5.20.0 is not parseable as XML.

Reproduces #533.