bugfix: pyTenable crashes if srcInterface or dstInterface not null
Addressed a bug causing a crash in when events->srcInterface or events->dstInterface was not null upon query completion.
The issue was traced back to incorrect usage of the marshmallow package.

Note: In Tenable.OT version 3.17, a bug existed where events->srcInterface and dstInterface consistently returned null.
Due to this behavior, the current bug did not manifest itself during that version

Additionally, added the retrieval of required nodes from dstInterface, ensuring that all necessary nodes are now properly obtained.